Started work on the actual menus of the app now. First the user will be presented with a menu that gives option of new project, load project and exit.
new project opens up a New Project window, that lets user set the project name. As this is done, it will generate a new binary file with the same name, which this project will save to. From here, an empty scene is loaded, and the user can start to add stuff to the project via menus. A save and exit option is available in the menu, as well as load sim data, load stl mesh, and probably some other options as well :)

load project will just load a project file from disk and set up the scene based on that.

exit will exit the application ;)